requirements-torch.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudpickle==1.3.0
contourpy==1.3.1
cycler==0.12.1
Cython==0.29.37
decorator==5.1.1
etils==1.11.0
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.3
fsspec==2024.10.0
future==1.0.0
gast==0.6.0
glfw==2.8.0
google-pasta==0.2.0
grpcio==1.68.1
-e git+https://github.com/froggagul/CSD-manipulation.git@526e17971ccfdc82207216114cb8a09e9bf25cc9#egg=gym&subdirectory=gym
gymnasium==1.0.0
gymnasium-robotics==1.3.1
h5py==3.12.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_resources==6.4.5
Jinja2==3.1.4
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml-dtypes==0.4.1
moviepy==2.1.1
mpmath==1.3.0
mujoco==3.1.6
mujoco-py==2.1.2.14
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pettingzoo==1.24.3
pillow==10.4.0
proglog==0.1.10
protobuf==5.29.1
pycparser==2.22
pyglet==1.5.0
Pygments==2.18.0
PyOpenGL==3.1.7
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rich==13.9.4
scipy==1.14.1
six==1.17.0
stable_baselines3==2.4.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
torch==2.5.1
torch-tb-profiler==0.4.3
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.21.0
